split ast into smaller files
