+++ ast/fragment.go
@@ -1,0 +1,22 @@
+package ast
+
+type FragmentSpread struct {
+	Name       string
+	Directives []Directive
+}
+
+type InlineFragment struct {
+	TypeCondition NamedType
+	Directives    []Directive
+	SelectionSet  SelectionSet
+}
+
+type FragmentDefinition struct {
+	Name string
+	// Note: fragment variable definitions are experimental and may be changed
+	// or removed in the future.
+	VariableDefinition []VariableDefinition
+	TypeCondition      NamedType
+	Directives         []Directive
+	SelectionSet       SelectionSet
+}
