+++ ast/selection.go
@@ -1,0 +1,24 @@
+package ast
+
+type SelectionSet []Selection
+
+type Selection interface {
+	isSelection()
+}
+
+func (Field) isSelection()          {}
+func (FragmentSpread) isSelection() {}
+func (InlineFragment) isSelection() {}
+
+type Field struct {
+	Alias        string
+	Name         string
+	Arguments    []Argument
+	Directives   []Directive
+	SelectionSet SelectionSet
+}
+
+type Argument struct {
+	Name  string
+	Value Value
+}
